Move demo_player.{h|c}pp to frontend
